fix: #60

+++ xposed/src/main/java/moe/fuqiuluo/shamrock/remote/service/data/Message.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.json.JsonElement
 @Serializable
 internal data class MessageResult(
     @SerialName("message_id") val msgId: Int,
-    @SerialName("time") val time: Double
+    @SerialName("time") val time: Long
 )
 
 @Serializable
